updated requested changes
