+++ src/AritfactoryUploader.UTest/JfrogRepoUpdaterTest.cs
@@ -41,97 +41,7 @@ namespace AritfactoryUploader.UTest
             // Assert
             Assert.That(actualAqlResultList.Count, Is.GreaterThan(2));
         }
-
-        [Test]
-        [TestCase("NPM", ".tgz")]
-        [TestCase("NUGET", ".nupkg")]
-        [TestCase("MAVEN", ".jar")]
-        [TestCase("DEBIAN", ".deb")]
-        [TestCase("POETRY", ".whl")]
-        [TestCase("CONAN", "package.tgz")]
-        public void GetPkgeNameExtensionBasedOnComponentType_GivenType_ReturnsPkgNameExtension(string type, string extension)
-        {
-            // Arrange
-            var package = new ComponentsToArtifactory();
-            package.ComponentType = type;
-            // Act
-            var actualExtension = JfrogRepoUpdater.GetPackageNameExtensionBasedOnComponentType(package);
-            // Assert
-            Assert.AreEqual(extension, actualExtension);
-        }
-        [Test]
-        public void GetUploadPackageDetails_CoversAllScenarios()
-        {
-            // Arrange
-            DisplayPackagesInfo displayPackagesInfo = new DisplayPackagesInfo()
-            {
-                JfrogFoundPackagesConan = new List<ComponentsToArtifactory>()
-                {
-                    new ComponentsToArtifactory()
-                    {
-                        ResponseMessage = new HttpResponseMessage()
-                        {
-                            StatusCode = HttpStatusCode.OK
-                        }
-                    }
-                },
-                JfrogFoundPackagesMaven = new List<ComponentsToArtifactory>()
-                {
-                    new ComponentsToArtifactory()
-                    {
-                        ResponseMessage = new HttpResponseMessage()
-                        {
-                            StatusCode = HttpStatusCode.OK
-                        }
-                    }
-                },
-                JfrogFoundPackagesNpm = new List<ComponentsToArtifactory>()
-                {
-                    new ComponentsToArtifactory()
-                    {
-                        ResponseMessage = new HttpResponseMessage()
-                        {
-                            StatusCode = HttpStatusCode.OK
-                        }
-                    }
-                },
-                JfrogFoundPackagesNuget = new List<ComponentsToArtifactory>()
-                {
-                    new ComponentsToArtifactory()
-                    {
-                        ResponseMessage = new HttpResponseMessage()
-                        {
-                            StatusCode = HttpStatusCode.OK
-                        }
-                    }
-                },
-                JfrogFoundPackagesPython = new List<ComponentsToArtifactory>()
-                {
-                    new ComponentsToArtifactory()
-                    {
-                        ResponseMessage = new HttpResponseMessage()
-                        {
-                            StatusCode = HttpStatusCode.OK
-                        }
-                    }
-                },
-                JfrogFoundPackagesDebian = new List<ComponentsToArtifactory>()
-                {
-                    new ComponentsToArtifactory()
-                    {
-                        ResponseMessage = new HttpResponseMessage()
-                        {
-                            StatusCode = HttpStatusCode.OK
-                        }
-                    }
-                }
-            };
-
-            // Act
-            List<ComponentsToArtifactory> uploadedPackages = JfrogRepoUpdater.GetUploadePackageDetails(displayPackagesInfo);
-
-            // Assert
-            Assert.AreEqual(6, uploadedPackages.Count);
-        }
+       
+        
     }
 }
